requirements.txt
Django==2.0.6
django-mysql==3.1.0
mysqlclient==1.4.2.post1
pkg-resources==0.0.0
pytz==2019.1
django-betterforms
